refactors signing flows to use a common helper hook
